[numerics] simplify VI tests
