+++ cmake/tools4tests.cmake
@@ -533,6 +533,35 @@ MACRO(NEW_RELAY_TEST)
   NEW_PB_TEST(relaytest ${ARGV})
 ENDMACRO(NEW_RELAY_TEST)
 
+MACRO(NEW_RELAY_TEST_1)
+
+  # check input file name
+  assert(SOURCE_FILE_NAME)
+  assert(TEST_NAME_PREFIX)
+  
+  STRING(CONCAT TEST_NAME ${TEST_NAME_PREFIX} "_" ${TEST_COLLECTION})
+
+  STRING(TOLOWER ${TEST_NAME} TEST_NAME)
+  CONFIGURE_FILE(${CMAKE_SOURCE_DIR}/numerics/src/RELAY/test/${SOURCE_FILE_NAME} 
+    ${CMAKE_CURRENT_BINARY_DIR}/${_CURRENT_TEST_DIRECTORY}/${TEST_NAME}.c)
+
+  SET(${TEST_NAME}_FSOURCES)
+
+  LIST(APPEND _EXE_LIST_${_CURRENT_TEST_DIRECTORY} ${TEST_NAME})
+  LIST(APPEND ${TEST_NAME}_FSOURCES ${CMAKE_CURRENT_BINARY_DIR}/${_CURRENT_TEST_DIRECTORY}/${TEST_NAME}.c)
+
+ENDMACRO(NEW_RELAY_TEST_1)
+
+MACRO(NEW_RELAY_TEST_COLLECTION)
+  # Set name of the file used to generate tests (c)source files.
+  set(SOURCE_FILE_NAME relay_test_collection.c.in )
+  set(TEST_NAME_PREFIX relay)
+  set(TEST_COLLECTION ${ARGV0})
+  NEW_RELAY_TEST_1(${ARGV})
+  unset(SOURCE_FILE_NAME)
+  unset(TEST_NAME_PREFIX)
+endmacro()
+
 MACRO(NEW_NCP_TEST)
   SET(FILE_TO_CONF ${ARGV0})
   SET(TEST_SOLVER ${ARGV1})
